refactor(eval-workflows): 归一维度绑定稳定性校验
