+++ src/eval-workflows/measurement/analysis/dimension-table-v1.ts
@@ -1,6 +1,7 @@
 import { z } from 'zod';
 import { IdentifierSchema, SamplingUnitIdsSchema, Sha256DigestSchema, canonicalizeJson, digestCanonicalJson, type JsonValue } from '../../../eval-core/contracts/index.js';
 import { analysisJsonSchema, analysisSchemaIdentity, compareStrings, round } from './analysis-support.js';
+import { assertStableBinding } from './dimension-binding.js';
 
 export const DIMENSION_TABLE_SCHEMA_VERSION = 'omk.dimension-table/v1' as const;
 export const DIMENSION_SCORE_DECIMALS = 2;
@@ -143,27 +144,6 @@ export function dimensionGroupId(group: Omit<DimensionGroup, 'groupId'>): string
       sourceGroupId: entry.sourceGroupId,
     })),
   });
-}
-
-function assertStableBinding(
-  entries: readonly DimensionEntry[],
-  keyField: 'dimensionId' | 'metricId' | 'sourceAnalysisResultId',
-  issue: Issue,
-): void {
-  const bindings = new Map<string, string>();
-  for (const entry of entries) {
-    const binding = canonicalizeJson({
-      dimensionId: entry.dimensionId,
-      metricId: entry.metricId,
-      sourceAnalysisResultId: entry.sourceAnalysisResultId,
-    });
-    const previous = bindings.get(entry[keyField]);
-    if (previous !== undefined && previous !== binding) {
-      issue(['groups'], `Dimension ${keyField} binding must remain stable across groups.`);
-      return;
-    }
-    bindings.set(entry[keyField], binding);
-  }
 }
 
 function validateDimensionTable(value: DimensionTableValue, issue: Issue): void {
